+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Implement `hi18n sync --check` (shorthand: `-c`) option to raise an error when files would be changed.
   It is useful when you want to ensure synchronization in your CI.
 - Switched command line parser (yargs to commander). The behavior may slightly change.
+- Fix `TypeError: Cannot read properties of undefined (reading 'node')`
+  on an empty Vocabulary or an empty Catalog in internally used rules.
 
 ## 0.1.4