+++ packages/cli/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.1.5
 
 - Implement `hi18n sync --check` (shorthand: `-c`) option to raise an error when files would be changed.
   It is useful when you want to ensure synchronization in your CI.